完成basic所有的接口 但feign调用未测试 todo待完成 socialize实现post接口

+++ basic/src/main/java/org/example/basic/service/UserService.java
@@ -14,7 +14,7 @@ public interface UserService extends IService<User> {
 
     User Login(String username, String password);
 
-    UserDto getUserById(int userId, int followId);
+    UserDto getUserById(long userId, long followId);
 
 
 }
